rename GeoData into geoimport
