+++ freecad/trails/geoimport/TestGeodat.py
