+++ freecad/trails/geoimport/gen_heightmap.py
