+++ freecad/trails/geoimport/geodat_lib.py
