+++ freecad/trails/geoimport/import_emir.py
