+++ freecad/trails/geoimport/import_image.py
