+++ freecad/trails/geoimport/import_latlony.py
