+++ freecad/trails/geoimport/import_lidar.py
