+++ freecad/trails/geoimport/import_osm.py
